requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
duckduckgo-search==4.4.3
python-docx==1.1.0
reportlab==4.0.9
PyPDF2==3.0.1
plotly==5.18.0
pandas==2.2.0